Move UsageAnonymizable newtypes in a separate file
